+++ src/components/common/InputBox.js
@@ -1,4 +1,4 @@
-import Component from '../core/Component.js';
+import Component from '../../core/Component.js';
 
 class InputBox extends Component {
   template() {
