+++ src/components/index.js
@@ -1,6 +1,6 @@
-export { default as Button } from './Button.js';
+export { default as Button } from './common/Button.js';
 export { default as Header } from './Header.js';
-export { default as InputBox } from './InputBox.js';
+export { default as InputBox } from './common/InputBox.js';
 export { default as InputModal } from './InputModal.js';
 export { default as Restaurant } from './Restaurant.js';
 export { default as RestaurantList } from './RestaurantList.js';
